requirements.txt
pytorch == 1.2.0
librosa == 0.8.0
pandas == 1.1.1
nengolib == 0.5.2
nengo == 2.8.0
h5py == 2.10.0
fastprogress == 1.0.0
audioread == 2.1.8
scikit-learn == 0.21.3